cleanup

+++ src/capture/capture_config.h
@@ -108,8 +108,8 @@ struct capture_conf {
   char capture_interface[IFNAMSIZ];                           /**< The capture interface name (any - to capture on all interfaces) */
   bool promiscuous;                                           /**< Specifies whether the interface is to be put into promiscuous mode. If promiscuous param is non-zero, promiscuous mode will be set, otherwise it will not be set */
   bool immediate;                                             /**< Sets whether immediate mode should be set on a capture handle when the handle is activated. If immediate param is non-zero, immediate mode will be set, otherwise it will not be set. */
-  uint16_t buffer_timeout;                                    /**< Specifies the packet buffer timeout, as a non-negative value, in milliseconds. (See pcap(3PCAP) for an explanation of the packet buffer timeout.) */
-  uint16_t process_interval;                                  /**< Specifies the packet process interval, in milliseconds. */ 
+  uint32_t buffer_timeout;                                    /**< Specifies the packet buffer timeout, as a non-negative value, in milliseconds. (See pcap(3PCAP) for an explanation of the packet buffer timeout.) */
+  uint32_t process_interval;                                  /**< Specifies the packet process interval, in milliseconds. */ 
   char analyser[MAX_ANALYSER_NAME_SIZE];                      /**< Specifies the packet analyser engine. */ 
   bool file_write;                                            /**< Specifies wether the packets should be saved to file(s). */
   bool db_write;                                              /**< Specifies wether the packets should be saved in a sqlite db. */
